# docs/swagger.yaml
basePath: /api/v2
definitions:
...
  apimodels.MessageResponse:
    properties:
      message:
        type: string
    type: object
  apimodels.Notify:
    properties:
      eventId:
        type: integer
      id:
        type: integer
      message:
        type: string
      subject:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
...
  apimodels.PostNotify:
    properties:
      eventId:
        type: integer
      message:
        type: string
      subject:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
  apimodels.Room:
    properties:
      id:
        type: integer
      title:
        type: string
    type: object
...
  apimodels.SuccessPostNotify:
    properties:
      id:
        type: integer
    type: object
...
info:
  contact: {}
  description: Swagger API for Dreamit-based web labs
  termsOfService: http://swagger.io/terms/
  title: Dreamit Swagger API
  version: "2.0"
paths:
...
  /notifies:
    get:
      description: Get notifies collection filtered by query parameters
      parameters:
      - description: Tags to filter notifies, any, optional
        in: query
        items:
          type: string
        name: tags
        type: array
      - description: Subject to filter notifies, single-match, optional
        in: query
        name: subject
        type: string
      - description: Event to filter notifies for, obligatory
        in: query
        name: eventId
        required: true
        type: integer
      - description: token with the bearer started
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/apimodels.Notify'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Get notifies collection
      tags:
      - notify
    post:
      consumes:
      - application/json
      description: Create notify and get id
      parameters:
      - description: New notify to add to the system
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/apimodels.PostNotify'
      - description: token with the bearer started
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apimodels.SuccessPostNotify'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Create new notify
      tags:
      - notify
  /notifies/{id}:
    delete:
      consumes:
      - application/json
      description: Delete notify by id
      parameters:
      - description: Notify ID
        in: path
        name: id
        required: true
        type: integer
      - description: token with the bearer started
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete notify by id
      tags:
      - notify
    get:
      description: Get notify by id
      parameters:
      - description: Notify ID
        in: path
        name: id
        required: true
        type: integer
      - description: token with the bearer started
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apimodels.Notify'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
      security:
      - ApiKeyAuth: []
      summary: Get notify
      tags:
      - notify
  /rooms:
    get:
      description: Get all rooms available in the system for booking
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/apimodels.Room'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
      summary: Get all rooms
      tags:
      - room
  /rooms/{id}:
    get:
      description: Get room by id
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/apimodels.Room'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apimodels.MessageResponse'
      summary: Get room
      tags:
      - room
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"